fix: handling of ipfs addresses in available transports
and some refactoring into multiple files
